change -cache.lib to -default.lib to work with older kicads
kicad has changed the default schematic library for example resistor
size, so loading this repo in any but the latest kicad results in
missed connections on the schematic.  putting the contents of -cache.lib
as the first library in the project ensures that older kicads will get
the correct parts, for example for resistors.

http://www.kicad-pcb.org/display/KICAD/Migration+from+old+stable+to+latest